+++ src/module/dataforged.ts
@@ -116,8 +116,12 @@ const COMPENDIUM_KEY_MAP = {
   Encounters: 'starforgedencounters',
 }
 const MARKDOWN_LINK_RE = new RegExp('\\[(.*?)\\]\\((.*?)\\)', 'g')
+const DESCRIPTOR_FOCUS_RE = new RegExp('\\[⏵Descriptor \\+ Focus\\]\\(.*?\\)')
 
 function renderLinksInStr(text: any, idMap: { [key: string]: string }): any {
+  // Catch "Descriptor+Focus" and replace with two links
+  text = text.replace(DESCRIPTOR_FOCUS_RE, '⏵ [Descriptor](Starforged/Oracles/Core/Descriptor) + [Focus](Starforged/Oracles/Core/Focus)')
+
   return text.replace(MARKDOWN_LINK_RE, (match, text, url) => {
     const kind = url.split('/')[1]
     const compendiumKey = COMPENDIUM_KEY_MAP[kind]